bounty-hunter/requirements.txt
PyGithub>=1.59.0
anthropic>=0.18.0
python-dotenv>=1.0.0
asyncio
